+++ src/test/java/com/sign/application/usecase/PasskeyRegistrationUseCaseTest.java
@@ -7,6 +7,7 @@ import com.sign.application.repository.HandleGenerator;
 import com.sign.application.repository.PasskeyRepository;
 import com.sign.dto.PasskeyRegistrationResult;
 import com.sign.infrastructure.repository.HandleGeneratorImpl;
+import com.sign.infrastructure.repository.InMemoryPasskeyRepository;
 import com.yubico.webauthn.RelyingParty;
 import com.yubico.webauthn.data.AuthenticatorAttestationResponse;
 import com.yubico.webauthn.data.ByteArray;
